refactor: move interfaces de usuários para subpacote dedicado
- Altera local de interfaces de `core.adapters.interfaces` para `core.adapters.interfaces.usuario`.
- Atualiza casos de uso, controladores e imports para refletir a nova organização.
- Otimiza estrutura dos pacotes para melhor modularidade e organização.
