+++ src/main/java/com/adjt/chefmanagerapi/core/adapters/interfaces/usuario/DeletarUsuarioPorId.java
@@ -1,4 +1,4 @@
-package com.adjt.chefmanagerapi.core.adapters.interfaces;
+package com.adjt.chefmanagerapi.core.adapters.interfaces.usuario;
 
 import java.util.UUID;
