+++ src/main/java/com/adjt/chefmanagerapi/infrastructure/api/controller/UsuarioController.java
@@ -1,8 +1,8 @@
 package com.adjt.chefmanagerapi.infrastructure.api.controller;
 
 import com.adjt.chefmanagerapi.UsuarioApi;
+import com.adjt.chefmanagerapi.core.adapters.interfaces.usuario.*;
 import com.adjt.chefmanagerapi.core.domain.dtos.usuario.UsuarioOutput;
-import com.adjt.chefmanagerapi.core.adapters.interfaces.*;
 import com.adjt.chefmanagerapi.infrastructure.api.mapper.UsuarioApiMapper;
 import com.adjt.chefmanagerapi.model.*;
 import org.springframework.http.HttpStatus;
